Resolved #1131 

Assembly became a little faster.

 | before | after
---- | ---- | ----
dist | 9264ms | **8761ms**
specs | 30116ms | **29160ms**
tmpl-specs | 2930ms | **2579ms**
examples | 7289ms | **6778ms**